funkcje pobierania obliczeń wypłat na podstawie danych z bazki przeniesione z index.php do functions.php
